requirements.txt
pandas==1.5.3
numpy==1.24.3
requests==2.31.0
scikit-learn==1.2.2
deap==1.3.3
backtesting==0.3.3
matplotlib==3.7.1
ta-lib==0.4.26
python-dotenv==1.0.0
ccxt==3.0.74
websocket-client==1.5.1
asyncio==3.4.3
aiohttp==3.8.4
pyyaml==6.0
pandas-ta==0.3.14b0